Implement RetryV2 event handler
