+++ chains/evm/listener/eventHandlers/tss.go
@@ -7,18 +7,10 @@ import (
 	"context"
 	"fmt"
 	"math/big"
-	"strings"
 
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 
-	"github.com/ChainSafe/sygma-relayer/chains/evm/calls/consts"
-	"github.com/ChainSafe/sygma-relayer/relayer/transfer"
-	"github.com/ChainSafe/sygma-relayer/store"
-	"github.com/ethereum/go-ethereum/accounts/abi"
-	"github.com/sygmaprotocol/sygma-core/relayer/message"
-
-	"github.com/ChainSafe/sygma-relayer/chains/evm/calls/events"
 	"github.com/ChainSafe/sygma-relayer/comm"
 	"github.com/ChainSafe/sygma-relayer/comm/p2p"
 	"github.com/ChainSafe/sygma-relayer/topology"
@@ -28,143 +20,8 @@ import (
 	frostKeygen "github.com/ChainSafe/sygma-relayer/tss/frost/keygen"
 	frostResharing "github.com/ChainSafe/sygma-relayer/tss/frost/resharing"
 	"github.com/ethereum/go-ethereum/common"
-	ethTypes "github.com/ethereum/go-ethereum/core/types"
 	"github.com/libp2p/go-libp2p/core/host"
 )
-
-type EventListener interface {
-	FetchKeygenEvents(ctx context.Context, address common.Address, startBlock *big.Int, endBlock *big.Int) ([]ethTypes.Log, error)
-	FetchFrostKeygenEvents(ctx context.Context, address common.Address, startBlock *big.Int, endBlock *big.Int) ([]ethTypes.Log, error)
-	FetchRefreshEvents(ctx context.Context, address common.Address, startBlock *big.Int, endBlock *big.Int) ([]*events.Refresh, error)
-	FetchRetryEvents(ctx context.Context, contractAddress common.Address, startBlock *big.Int, endBlock *big.Int) ([]events.RetryEvent, error)
-	FetchRetryDepositEvents(event events.RetryEvent, bridgeAddress common.Address, blockConfirmations *big.Int) ([]events.Deposit, error)
-	FetchDeposits(ctx context.Context, address common.Address, startBlock *big.Int, endBlock *big.Int) ([]*events.Deposit, error)
-}
-
-type PropStorer interface {
-	StorePropStatus(source, destination uint8, depositNonce uint64, status store.PropStatus) error
-	PropStatus(source, destination uint8, depositNonce uint64) (store.PropStatus, error)
-}
-
-type RetryEventHandler struct {
-	log                zerolog.Logger
-	eventListener      EventListener
-	depositHandler     DepositHandler
-	propStorer         PropStorer
-	bridgeAddress      common.Address
-	bridgeABI          abi.ABI
-	domainID           uint8
-	blockConfirmations *big.Int
-	msgChan            chan []*message.Message
-}
-
-func NewRetryEventHandler(
-	logC zerolog.Context,
-	eventListener EventListener,
-	depositHandler DepositHandler,
-	propStorer PropStorer,
-	bridgeAddress common.Address,
-	domainID uint8,
-	blockConfirmations *big.Int,
-	msgChan chan []*message.Message,
-) *RetryEventHandler {
-	bridgeABI, _ := abi.JSON(strings.NewReader(consts.BridgeABI))
-	return &RetryEventHandler{
-		log:                logC.Logger(),
-		eventListener:      eventListener,
-		depositHandler:     depositHandler,
-		propStorer:         propStorer,
-		bridgeAddress:      bridgeAddress,
-		bridgeABI:          bridgeABI,
-		domainID:           domainID,
-		blockConfirmations: blockConfirmations,
-		msgChan:            msgChan,
-	}
-}
-
-func (eh *RetryEventHandler) HandleEvents(
-	startBlock *big.Int,
-	endBlock *big.Int,
-) error {
-	retryEvents, err := eh.eventListener.FetchRetryEvents(context.Background(), eh.bridgeAddress, startBlock, endBlock)
-	if err != nil {
-		return fmt.Errorf("unable to fetch retry events because of: %+v", err)
-	}
-
-	retriesByDomain := make(map[uint8][]*message.Message)
-	for _, event := range retryEvents {
-		func(event events.RetryEvent) {
-			defer func() {
-				if r := recover(); r != nil {
-					eh.log.Error().Err(err).Msgf("panic occured while handling retry event %+v", event)
-				}
-			}()
-
-			deposits, err := eh.eventListener.FetchRetryDepositEvents(event, eh.bridgeAddress, eh.blockConfirmations)
-			if err != nil {
-				eh.log.Error().Err(err).Msgf("Unable to fetch deposit events from event %+v", event)
-				return
-			}
-
-			for _, d := range deposits {
-				messageID := fmt.Sprintf("retry-%d-%d-%d-%d", eh.domainID, d.DestinationDomainID, startBlock, endBlock)
-				msg, err := eh.depositHandler.HandleDeposit(
-					eh.domainID, d.DestinationDomainID, d.DepositNonce,
-					d.ResourceID, d.Data, d.HandlerResponse, messageID,
-				)
-				if err != nil {
-					eh.log.Err(err).Str("messageID", msg.ID).Msgf("Failed handling deposit %+v", d)
-					continue
-				}
-				isExecuted, err := eh.isExecuted(msg)
-				if err != nil {
-					eh.log.Err(err).Str("messageID", msg.ID).Msgf("Failed checking if deposit executed %+v", d)
-					continue
-				}
-				if isExecuted {
-					eh.log.Debug().Str("messageID", msg.ID).Msgf("Deposit marked as executed %+v", d)
-					continue
-				}
-
-				eh.log.Info().Str("messageID", msg.ID).Msgf(
-					"Resolved retry message %+v in block range: %s-%s", msg, startBlock.String(), endBlock.String(),
-				)
-				retriesByDomain[msg.Destination] = append(retriesByDomain[msg.Destination], msg)
-			}
-		}(event)
-	}
-
-	for _, retries := range retriesByDomain {
-		eh.msgChan <- retries
-	}
-
-	return nil
-}
-
-func (eh *RetryEventHandler) isExecuted(msg *message.Message) (bool, error) {
-	var err error
-	propStatus, err := eh.propStorer.PropStatus(
-		msg.Source,
-		msg.Destination,
-		msg.Data.(transfer.TransferMessageData).DepositNonce)
-	if err != nil {
-		return true, err
-	}
-
-	if propStatus == store.ExecutedProp {
-		return true, nil
-	}
-
-	// change the status to failed if proposal is stuck to be able to retry it
-	if propStatus == store.PendingProp {
-		err = eh.propStorer.StorePropStatus(
-			msg.Source,
-			msg.Destination,
-			msg.Data.(transfer.TransferMessageData).DepositNonce,
-			store.FailedProp)
-	}
-	return false, err
-}
 
 type KeygenEventHandler struct {
 	log           zerolog.Logger
@@ -397,62 +254,4 @@ func (eh *RefreshEventHandler) HandleEvents(
 
 func (eh *RefreshEventHandler) sessionID(block *big.Int) string {
 	return fmt.Sprintf("resharing-%s", block.String())
-}
-
-type DepositHandler interface {
-	HandleDeposit(sourceID, destID uint8, nonce uint64, resourceID [32]byte, calldata, handlerResponse []byte, messageID string) (*message.Message, error)
-}
-
-type DepositEventHandler struct {
-	eventListener  EventListener
-	depositHandler DepositHandler
-	bridgeAddress  common.Address
-	domainID       uint8
-	msgChan        chan []*message.Message
-}
-
-func NewDepositEventHandler(eventListener EventListener, depositHandler DepositHandler, bridgeAddress common.Address, domainID uint8, msgChan chan []*message.Message) *DepositEventHandler {
-	return &DepositEventHandler{
-		eventListener:  eventListener,
-		depositHandler: depositHandler,
-		bridgeAddress:  bridgeAddress,
-		domainID:       domainID,
-		msgChan:        msgChan,
-	}
-}
-
-func (eh *DepositEventHandler) HandleEvents(startBlock *big.Int, endBlock *big.Int) error {
-	deposits, err := eh.eventListener.FetchDeposits(context.Background(), eh.bridgeAddress, startBlock, endBlock)
-	if err != nil {
-		return fmt.Errorf("unable to fetch deposit events because of: %+v", err)
-	}
-
-	domainDeposits := make(map[uint8][]*message.Message)
-	for _, d := range deposits {
-		func(d *events.Deposit) {
-			defer func() {
-				if r := recover(); r != nil {
-					log.Error().Err(err).Msgf("panic occured while handling deposit %+v", d)
-				}
-			}()
-
-			messageID := fmt.Sprintf("%d-%d-%d-%d", eh.domainID, d.DestinationDomainID, startBlock, endBlock)
-			m, err := eh.depositHandler.HandleDeposit(eh.domainID, d.DestinationDomainID, d.DepositNonce, d.ResourceID, d.Data, d.HandlerResponse, messageID)
-			if err != nil {
-				log.Error().Err(err).Str("start block", startBlock.String()).Str("end block", endBlock.String()).Uint8("domainID", eh.domainID).Msgf("%v", err)
-				return
-			}
-
-			log.Debug().Str("messageID", m.ID).Msgf("Resolved message %+v in block range: %s-%s", m, startBlock.String(), endBlock.String())
-			domainDeposits[m.Destination] = append(domainDeposits[m.Destination], m)
-		}(d)
-	}
-
-	for _, deposits := range domainDeposits {
-		go func(d []*message.Message) {
-			eh.msgChan <- d
-		}(deposits)
-	}
-
-	return nil
 }
